Add db preparations and a bit more related db code
The tile db initial init now uses a transaction, so generating it takes
lass than a second instead of multiple hours like before.
Yeah, that wasn't so difficult after all...

Another random mix of changes, still setting things up. Too lazy to
split these up into separate commits since a lot of the changes depend
on each other at this point.

Also set it up to use an in-memory db by default, so it doesn't crash
and burn when it doesn't find the prod db file where it expects it.