Implement basic rate limiting for tile placements
The params are compile-time tweakable, still looking for ones that work
nicely. It just rejects tile placements if the set rate is exceeded.
The currently set max rate is 7 placements/sec, which is quite high.
~8/sec is about as fast as I can click, so I should figure out a
realistic value someone might be clicking when actually drawing
something.

There are bugs. The rate limit allowance value currently isn't saved to
the database, so it gets reset when re-authenticating, giving bots a way
around the limit.
The client also falls out of sync with the server when rate limiting
happens. We need to add logic in the client to only update the progress
bar states if the tile placement was valid.