Sloppily slapped together code that migrated the 512px canvas to 1024px
I wanted the existing canvas shifted to the bottom left so the
SerenityOS taskbar would be in the right place. I couldn't figure out
JOINs in sqlite so I just put together this code to do it.
The rough workflow to use this is:

- Load up db in sqlite
- ALTER TABLE tiles RENAME TO tiles_old;
- Comment out canvas_save_timer_fn() call at the bottom of main.c
- Compile and run with params.json pointing to the desired db file
- It'll generated a new empty 1024px canvas.
- Load db again in sqlite console to add in indices
- CREATE INDEX tmp_ix ON tiles(X,Y);
- CREATE INDEX tmp_ix_2 ON tiles(Y,X);
- Uncomment canvas_save_timer_fn() in main.c and recompile
- Run bin/nmc again, it should migrate and save the new canvas.
- Open db file in sqliet console
- DROP TABLE tiles_old;
- DROP INDEX tmp_ix;
- DROP INDEX tmp_ix_2;
- CREATE INDEX tile_coord_ix ON tiles(X,Y);
- CREATE INDEX tile_coord_ix_2 ON tiles(Y,X);

And you should be done! Easy, right? Good enough for a one-off
migration.