Validate that every user is logged in only once at a time
I have a suspicion taht allowing a user to be logged in multiple times
at once has resulted in some silent data races and data integrity
issues, so I'm enforcing this now. If a user has already auth'd, kick
the other session. I also added support for this in the client, so it
presents a convenient 'reconnect here' button so you don't have to go
hunting for that other tab.