Implement an automatic kick for inactive users
Configurable in params.json. It'll now kick a user who hasn't sent any
request for an hour.
The check is done every users_save_interval_sec seconds, so it isn't
precie to the second, but it's nice enough for this feature.

This is slightly annoying for that use case where one likes to keep the
canvas open to monitor new drawings appear. I think I'll bump up the
idle timer even more, and then lower it in times of high traffic.