Remove global canvas state (g_canvas)
Big shuffle, now it's a stack local in main(), and gets dished out to
everyone that wants to modify it. Much cleaner.

Also moved config reloading and db backup handling out of the signal
handler (yikes!) to the main thread runloop, with global bools that
trigger them. I'm frankly surprised that that even worked as well as it
did!