client: Start rewriting bloated React client
The old React app is quite overbuilt for how small this thing is, so I'm
rewriting all the functionality into a single .html and .js.
So far, the source is far more readable, and I hope to keep it small
enough so I can just ship the unminified source to browsers.
Currently about ~70% complete, I also plan to drop some features that
didn't really pan out, such as levels.